requirements.txt
python-dotenv==1.0.0
tweepy==4.14.0
google-generativeai==0.3.2
python-dateutil==2.8.2 